Added unittests for the class 1) OutOfBoundaryEventListener 2) ActionRandomChangeDirection Fixed following, 1) Typecasting issue with ActionRandomDirectionChange. 2) Was verifying width, instead of height, for the horizontal floor of gamepanel.

+++ client/src/main/java/action/ActionChangeSpeed.java
@@ -6,9 +6,9 @@ import model.SpriteModel;
 
 public class ActionChangeSpeed implements GameAction,Serializable {
 
-    private int newSpeedX, newSpeedY;
+    private double newSpeedX, newSpeedY;
 
-    public ActionChangeSpeed(int newSpeedX, int newSpeedY) {
+    public ActionChangeSpeed(double newSpeedX, double newSpeedY) {
         super();
         this.newSpeedX = newSpeedX;
         this.newSpeedY = newSpeedY;
